Fix corrupt json
Tested on kibana locally this import succeeds and all the visualisations
appear to work as expected.

It's difficult to know if these are the latest versions of the
visualisations as the authors intended due to the complexity of the
nested json. As such this is more of a "best guess" of what the json
should be.